update to accept file, message, or both

+++ backend/controllers/comments.js
@@ -5,14 +5,22 @@ const Like = db.like;
 const fs = require('fs');
 
 exports.create = (req, res, next) => {
-  Comment.create({
-		message: req.body.message,
-    attachement: `${req.protocol}://${req.get('host')}/images/comments/${req.file.filename}`,
+	const commentObject = req.file ? // Contrôle si req.file existe
+	{
+		...req.body,
+		attachement: `${req.protocol}://${req.get('host')}/images/comments/${req.file.filename}`,
 		like: 0,
 		dislike: 0,
 		publicationId: req.body.publicationId,
 		userId: userId(req)
-  })
+	} : { 
+		...req.body,
+		like: 0,
+		dislike: 0,
+		publicationId: req.body.publicationId,
+		userId: userId(req)
+	}
+  Comment.create( commentObject )
     .then(() => res.status(201).json({ message: 'Commentaire ajouté !' }))
     .catch(err => res.status(500).send({ message: err.message }));
 };
@@ -27,8 +35,10 @@ exports.modify = (req, res, next) => {
 	.then(comment => {
 		if (comment.userId == userId(req)) {
 			if (req.file != undefined) {
-				const filename = comment.attachement.split('/images/comments/')[1];
-				fs.unlinkSync(`images/comments/${filename}`)
+				if (comment.attachement != null) {
+					const filename = comment.attachement.split('/images/comments/')[1];
+					fs.unlinkSync(`images/comments/${filename}`)
+				}
 			}
 			Comment.update( commentObject, { where: { id: req.params.id }})
 				.then(() => res.status(200).json({ message: 'Commentaire modifiée !' }))
@@ -45,8 +55,8 @@ exports.delete = (req, res, next) => {
 	Comment.findOne({ where: { id: req.params.id }})
 		.then(comment => {
 			if (comment.userId == userId(req)) {
-				const filename = comment.attachement.split('/images/comments/')[1];
-				if (filename != undefined) {
+				if (comment.attachement != null) {
+					const filename = comment.attachement.split('/images/comments/')[1];
 					fs.unlinkSync(`images/comments/${filename}`)
 				}
 				Comment.destroy({	where: { id: req.params.id }})
